requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.0
pydantic>=2.5.0
python-dotenv>=1.0.0
cachetools>=5.3.0
